+++ src/app/shell/admin-tools/platform/platform-info/info-edit/info-edit/info-form.component.scss
