+++ src/app/shell/admin-tools/platform/platform-info/info-edit/info-edit/info-form.component.spec.ts
@@ -3,7 +3,7 @@ import { MatFormFieldModule } from '@angular/material/form-field';
 import { FormControl, FormGroup, ReactiveFormsModule } from '@angular/forms';
 import { MatInputModule } from '@angular/material/input';
 import { MatIconModule } from '@angular/material/icon';
-import { InfoFormComponent } from './platform-info-form.component';
+import { InfoFormComponent } from './info-form.component';
 import { Component, Input } from '@angular/core';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
